Add the new servlex-loader project.
Its purpose is to provide one class, which is a class loader to be set
as a Tomcat webapp class loader (that is, Servlex class loader) so it
looks into the JAR files in the package repository (extensions for Saxon
and Calabash can use JAR files).

It has to be a separate project, as it has to be a separate JAR file, as
it has to be copied into Tomcat's own lib/ directory.