Do not use sendError().
We control the error page generation, we do not use the error page
generation system from the Servlet spec.  So use setStatus() instead,
and simply write the error page explicitly in the output stream.

Untill this change, in case of a sendError(500, ...), it was using the
default HTML Tomcat page for 500, even though we were writing a nice XML
error message to the output stream.