Do not relly on Calabash to run the tests.
Calabash still relies on old versions of the Apache HTTP Client, not
compatible with the latest versions of the EXPath HTTP Client.  So this
removes the dependency on it, but compiling, running and formatting
XSpec test suites "by hand" (and adding a dependengy on the path where
XSpec is installed).

To be revert back when Calabash add proper support.